pyproject.toml
[tool.poetry]
name = "used_cars_api"
version = "0.1.0"
description = "An API to serve Cragslist used cars for the recommendation service"
authors = ["nik vostrosablin (nikita.vostrosablin@gmail.com)"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "0.23.2"
pydantic = "^2.0.0"
requests = "^2.22.0"
loguru = "^0.7.0"
joblib = "^1.2.0"
asyncpg = "^0.28.0"
greenlet = "^2.0.2"

[tool.poetry.group.database.dependencies]
pandas = "^2.1.0"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
kaggle = "^1.6.0a3"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
black = {version = "^19.10b0", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^2.15.8"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''